+++ .eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-	extends: '@eslint-sets/ts',
-	rules: {
-		'no-only-tests/no-only-tests': 0
-	}
-}