+++ scripts/sync.ts
@@ -1,5 +1,5 @@
-import { execSync } from 'child_process'
+import { execSync } from 'node:child_process'
 
 execSync(
-	'curl -X PUT -d "sync_upstream=true" "https://registry-direct.npmmirror.com/vue-mount-plugin/sync"'
+	'curl -X PUT -d "sync_upstream=true" "https://registry-direct.npmmirror.com/vue-mount-plugin/sync"',
 )