+++ src/__tests__/index.spec.ts
@@ -1,3 +1,0 @@
-test('defines init() & create()', () => {
-	//
-})