+++ src/utils/vue.ts
@@ -1,16 +1,13 @@
 import type { App, AppContext, Plugin } from 'vue-demi'
 
-export type SFCWithInstall<T> = T &
-	Plugin & {
-		name: string
-	}
+export type SFCWithInstall<T> = T & Plugin & { name: string }
 export type SFCInstallWithContext<T> = SFCWithInstall<T> & {
 	_context: AppContext | null
 }
 
 export const noop = (): void => {}
 
-export function withInstall<T, E extends Record<string, unknown>>(component: T, extra?: E) {
+export function withInstall<T, E extends Record<string, unknown>>(component: T, extra?: E): SFCWithInstall<T> {
 	/* c8 ignore next */
 	;(component as SFCWithInstall<T>).install = (app: App): void => {
 		for (const comp of [component, ...Object.values(extra ?? {})] as SFCWithInstall<T>[]) {
@@ -22,20 +19,23 @@ export function withInstall<T, E extends Record<string, unknown>>(component: T, 
 			;(component as any)[key] = comp
 		}
 	}
+
 	return component as SFCWithInstall<T>
 }
 
-export function withFunctionInstall<T>(component: T, name: string) {
+export function withFunctionInstall<T>(component: T, name: string): SFCInstallWithContext<T> {
 	/* c8 ignore next */
 	;(component as SFCInstallWithContext<T>).install = (app: App): void => {
 		;(component as SFCInstallWithContext<T>)._context = app._context
 		app.config.globalProperties[name] = component
 	}
+
 	return component as SFCInstallWithContext<T>
 }
 
-export function withNoopInstall<T>(component: T) {
+export function withNoopInstall<T>(component: T): SFCWithInstall<T> {
 	/* c8 ignore next */
 	;(component as SFCWithInstall<T>).install = noop
+
 	return component as SFCWithInstall<T>
 }